Remove GetFinalPathNameByHandle calls
GetFinalPathNameByHandle was added in vista so cannot be used on xp, it has been poorly replaced with NtQueryInformationFile which is less picky about the form it returns the path in as it for example doesn't normalise the path.
This commit also fails to build due to not removing a call from some of the python code.