Hardcode uname output on windows
For some reason make errors with code 587202932, returns control to windows which causes the shell to show the prompt but then builds as normal.
I thought I was imagining it at first but windows builds are definitely way faster now. Not sure if it beats wsl let alone a regular linux install but its way more usable now.