Remove unused code
The windows tramp never worked right and the main branch
just got rid of it so decided to revert most of my local
changes as well. I have kept cosmo_dlprep in case I need
it in future for getting gtk3 working. Last time I tried
I had issues when the swapped callback being the same as
the regular one and I have no clue if that still happens
This doesn't change anything with my raylib code as that
already used GetProcAddress and ms_abi to get around the
parameter issues which still works but I can instead use
cosmo_dlsym as it is just a wrapper with error checking.